major change to use synchronous pull
Asynchronous pulling (using futures and "subscribe") uses
a buffer in the gRPC stream. So, messages enter that buffer
and stay there. When we have long-running processes with each
message and many subscribers, messages can appear to be "blocked"
in the pubsub queue. The way around this is to handle all the
pubsub pulling synchronously. This also necessitates resetting the
ack deadline regularly. All good, though, and actually more
transparent.